Move Article to Articles/Article (revert change)
